Separating response handling from Code module class
